Improve data fetching error display and redirect to landing page after changing the display settings (#58)
When data fetching for tiles in the landing page fails, an error notification is rendered in place of the whole tile, instead of within the tile. Also, the user is not redirected to the landing page after changing which tiles should be displayed on the settings page. This behaviour is common in the SCM-Manager. Both of these issues have been fixed in this PR.