add handling for deleted repositories (#65)
When a repository is deleted, it should be displayed as an event in the activities tile. Also, when a repository is deleted/renamed, previous events regarding this repository should have their title text stroked through. This is achieved by manually going through all stored events and marking the relevant ones as "deleted".

Bulma's styling interacts unfavorably with the core change, which required linkless event tiles to be wrapped in a non-functional div to avoid mismatching visuals.

Co-authored-by: Eduard Heimbuch <eduard.heimbuch@cloudogu.com>