Fixes an error after deleting a favorite repository
After deleting a repository, it could take a little while
until the deleted repository is removed from the favorite
lists for all users. During this time, loading the
favorites leads to a NullPointerException in the backend
(or an "Cannot read properties of null" error in the
frontend for older versions of the plugin), because
the favorite service still tries to read this repository
and does not handle the null result correctly.

Committed-by: Eduard Heimbuch <eduard.heimbuch@cloudogu.com>
Co-authored-by: René Pfeuffer <rene.pfeuffer@cloudogu.com>